Bricks Visible
Brick locations are currently hard coded in, and one is where the ball spawns. Variable "invisBrick" set to true invokes a bug I found where the bricks are only visible for a brief moment (timer value probably) when hit.

To do:
Make ball bounce up or down when hitting the top or bottom of a brick.
Change brick generator from hard coded.